ml/requirements.txt
# ML Requirements
spacy==3.4.4
scikit-learn
pandas==2.0.3
numpy==1.24.3
PyPDF2==3.0.1
python-docx==0.8.11
matplotlib==3.7.1
seaborn==0.12.2